+++ src/SomeDL/core/downloader.py
@@ -10,13 +10,13 @@ from SomeDL.utils.utils import sanitize
 from SomeDL.utils.utils import generateOutputName
 
 
-def downloadSong(videoID: str, artist: str, song: str, album, date, track_pos, track_count): 
+def downloadSong(videoID: str, artist: str, album_artist: str, song: str, album, date, track_pos, track_count):
     # https://github.com/yt-dlp/yt-dlp?tab=readme-ov-file#embedding-yt-dlp. 
     #URLS = ['https://music.youtube.com/watch?v=hComisqDS1I']
     URLS = f'https://music.youtube.com/watch?v={videoID}'
     #URLS = f'https://www.youtube.com/watch?v=-X8Olge799M'
     # output_template = sanitize(f'{artist} - {song}')
-    output_template = generateOutputName(artist, song, album, date, track_pos, track_count)
+    output_template = generateOutputName(artist, album_artist, song, album, date, track_pos, track_count)
 
     log.info("Start yt-dlp download...")